+++ src/Api/ConfigInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace JosephLeedy\CustomFees\Api;
 
+use JosephLeedy\CustomFees\Model\FeeStatus;
 use JosephLeedy\CustomFees\Model\FeeType;
 use Magento\Framework\Exception\LocalizedException;
 
@@ -17,6 +18,7 @@ interface ConfigInterface
      *     title: string,
      *     type: value-of<FeeType>,
      *     value: float,
+     *     status: value-of<FeeStatus>,
      *     advanced: array{
      *         conditions?: array{
      *             type: class-string,